backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
llama-index==0.9.48
llama-index-llms-openai
llama-index-embeddings-openai
llama-index-vector-stores-weaviate
weaviate-client==3.25.3
openai==1.12.0
pdfplumber==0.10.3
PyPDF2==3.0.1
python-docx==1.1.0
beautifulsoup4==4.12.3
redis==5.0.1
motor==3.3.2
python-multipart==0.0.6
aiofiles==23.2.1
